// resources/js/api/generated/models/volume.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { VolumeDriverOpts } from './volumeDriverOpts';
import type { VolumeLabels } from './volumeLabels';
import type { VolumeUsage } from './volumeUsage';

export interface Volume {
  created?: string;
  driver?: string;
  driver_opts?: VolumeDriverOpts;
  exists: boolean;
  external?: boolean;
  labels?: VolumeLabels;
  mountpoint?: string;
  name: string;
  scope?: string;
  used_by?: VolumeUsage[];
}